Cleanup helper methods of the revision reconciler.
- Remove a build helper, which is no longer used.
- Move the remaining helper to it's only callsite to remove the `helpers.go` file.
