--- pkg/reconciler/revision/helpers.go
@@ -1,50 +1,0 @@
-/*
-Copyright 2018 The Knative Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package revision
-
-import (
-	duckv1alpha1 "github.com/knative/pkg/apis/duck/v1alpha1"
-	appsv1 "k8s.io/api/apps/v1"
-	corev1 "k8s.io/api/core/v1"
-)
-
-// TODO(mattmoor): This should be a helper on Build (upstream)
-func getBuildDoneCondition(build *duckv1alpha1.KResource) *duckv1alpha1.Condition {
-	for _, cond := range build.Status.Conditions {
-		if cond.Status == corev1.ConditionUnknown {
-			continue
-		}
-		return &cond
-	}
-	return nil
-}
-
-func hasDeploymentTimedOut(deployment *appsv1.Deployment) bool {
-	// as per https://kubernetes.io/docs/concepts/workloads/controllers/deployment
-	for _, cond := range deployment.Status.Conditions {
-		// Look for Deployment with status False
-		if cond.Status != corev1.ConditionFalse {
-			continue
-		}
-		// with Type Progressing and Reason Timeout
-		// TODO(arvtiwar): hard coding "ProgressDeadlineExceeded" to avoid import kubernetes/kubernetes
-		if cond.Type == appsv1.DeploymentProgressing && cond.Reason == "ProgressDeadlineExceeded" {
-			return true
-		}
-	}
-	return false
-}
